requirements.txt
Flask==2.0.2
numpy==1.21.2
matplotlib==3.4.3
tensorflow==2.6.0
Pillow==8.3.2
